Feat(newMain): Added UpperCase.cs and did major refactor.
