+++ Caseify.csproj
@@ -46,6 +46,7 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="Commands\CamelCase.cs" />
+    <Compile Include="Commands\UpperCase.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Commands\TitleCase.cs" />
     <Compile Include="CaseifyPackage.cs" />
