+++ Commands/CamelCase.cs
@@ -8,15 +8,7 @@ internal sealed class CamelCase : BaseCommand<CamelCase>
 {
     protected override async Task ExecuteAsync(OleMenuCmdEventArgs eventArgs)
     {
-        var docView = await VS.Documents.GetActiveDocumentViewAsync().ConfigureAwait(false);
-        var selection = docView?.TextView?.Selection.SelectedSpans.FirstOrDefault();
-
-        if (selection.HasValue)
-        {
-            var selectedText = selection.GetValueOrDefault().GetText();
-            var camelCased = ConvertToCamelCase(selectedText);
-            docView.TextView.TextBuffer.Replace(selection.Value, camelCased);
-        }
+        await Transform.TextConverterAsync(ConvertToCamelCase);
     }
 
     private static string ConvertToCamelCase(string selectedText)
