+++ Commands/TitleCase.cs
@@ -1,5 +1,5 @@
 ﻿using System.Globalization;
-using System.Linq;
+using Caseify.Utility;
 
 namespace Caseify.Commands
 {
@@ -8,15 +8,7 @@ namespace Caseify.Commands
     {
         protected override async Task ExecuteAsync(OleMenuCmdEventArgs e)
         {
-            var docView = await VS.Documents.GetActiveDocumentViewAsync().ConfigureAwait(false);
-            var selection = docView?.TextView?.Selection.SelectedSpans.FirstOrDefault();
-
-            if (selection.HasValue)
-            {
-                var selectedText = selection.GetValueOrDefault().GetText();
-                var transformedText = CultureInfo.CurrentCulture.TextInfo.ToTitleCase(selectedText);
-                docView.TextView.TextBuffer.Replace(selection.Value, transformedText);
-            }
+            await Transform.TextConverterAsync(CultureInfo.CurrentCulture.TextInfo.ToTitleCase);
         }
     }
 }
