+++ source.extension.cs
@@ -11,7 +11,7 @@ namespace Caseify
         public const string Name = "Caseify";
         public const string Description = @"Transforming or modifying the casing of code elements.";
         public const string Language = "en-US";
-        public const string Version = "1.0";
+        public const string Version = "1.1";
         public const string Author = "Sailesh B K";
         public const string Tags = "Caseify";
     }
